Revert "Fix compiler warning."
This reverts commit 83ee04e4e3f108320286213f37d394d06117448c.

+++ trunk/src/jsformatter.cpp
@@ -80,7 +80,7 @@ DLLAPI const char *JSFGetDebugOutput(JSFormatter *jsf)
 	return ((RealJSFormatter *)jsf)->GetDebugOutput();
 }
 
-DLLAPI unsigned int JSFGetFormattedLine(JSFormatter *jsf, unsigned int originalLine)
+DLLAPI long JSFGetFormattedLine(JSFormatter *jsf, long originalLine)
 {
 	return ((RealJSFormatter *)jsf)->GetFormattedLine(originalLine);
 }
